chore: Revamp store to have operation specific files
